+++ Core/include/structs.h
